restructured application and added database models, CAS functionality, and fixed broken links.
