+++ .gitignore
@@ -1,3 +1,4 @@
 env/*
 .vscode/*
 .DS_Store
+config.py
