+++ teambuildingapp/static/team_manager_panel.css
