+++ teambuildingapp/templates/team_manager_panel.html
